Setup C app to use OpenSSL-Universal with xCode & macOS
## Setup a C app to use OpenSSL-Universal with xCode and macOS.
This tutorial assumes you want to:   1) Write a C application.  2) Use OpenSSL libraries for some crypto operation.  3) Use a Mac and xCode 8.  4) Use CocoaPods and the OpenSSL-Universal pod as an elegant way to stay updated with the latest libraries and headers.    
#####Create your macOS HelloWorld C app
If you have not done this before with xCode, select  `File / New Project / macOS / Command Line Tool`.  You just created a default Hello World C project.
#####Setup Cocoa Pods
Make sure your machine is setup for CocoaPods.  If not, follow this [great tutorial].    After CocoaPods is setup, open `Terminal` and navigate to your project folder and run `pod init`.  After that has completed, type `open -a Xcode Podfile`.   
#####Get the OpenSSL-Universal pod
Add `pod 'OpenSSL-Universal'` inside the podfile. Save the file.  Return to `Terminal` and run `Pod Install`.      After that download completes, make sure to close the C project and open the workspace file that was created. 
#####What is inside of the OpenSSL-Universal pod?
Once this successfully completed, you have now got access to a pre-compiled version of the static OpenSSL libraries and the C header files required to call these functions.  
#####Where can I use OpenSSL-Universal?
These libraries work for both iOS and MacOS.  There are two OpenSSL static libraries; `libcrypto.a` and `libssl.a`     ::Do NOT expect these OpenSSL files to work on every CPU architecture in the world. It is your prerogative to check.  Ask yourself, are you trying to write an app for old devices? new devices only? all iOS devices? only macOS?, etc ::
##### Fat Binaries
OpenSSL-Universal libraries are referred to as Fat Binaries as they support multiple CPU architectures.  For example, return to `Terminal`, navigate to your OpenSSL-Universal macOS files and run the command `file libcrypto.a`  This will tell you architecture the file is compiled against `x86_64` and `x386` .
##### xCode Setup
You want to ensure xCode knows; 

1. Where the OpenSSL static libraries are located.
2. Where the OpenSSL header files are located for the C include statements. 
Inside your workspace, go to the Target (not the Project).  The Target is the C app that is produced after a successful build.  Select `Build Phases` and `Link Binary With Libraries`.    Select `+` and navigate to the static OpenSSL libraries that are included in the framework.  Now go to the Project (not the Target).  In `Build Settings` set the `Always Search User Paths` to `Yes`.  Then add a new entry to the `User Header Search Paths`. This should be the location of the OpenSSL header files that were included in OpenSSLUniversal.

##### Test your app build: include the OpenSSL Header file
'' include "openssl/evp.h" 
'' void generateSHA256Hash(const char *ptr)
'' {
	 '' some nice C code 
'' }
	''