Update bundle version numbers
It turns out that AppStore requires bundles to have strictly three-part
semver-style versions. Replace "1.0.2.14" with "1.0.214" in the short
identifier. Note that we have to keep the patch version in there because
this is how the OS will decide on updates, so we have to keep it
different and at least somewhat ordered.