Correct umbrella header compilation
It turns out that we cannot simply take all the headers and include them
from the umbrella header. OpenSSL has some intricate interdependencies
there so we have to resort to hacks.

Instead of using "find" to locate all the headers (which returns them
in some unspecified order), make sure to include the "ssl.h" header
first and then follow it with other headers in alphabetic order.
This seems to work for the current OpenSSL version (1.0.2o).

Take care to not include the umbrella header into itself to avoid
blowing up compiler's include stack.