Add support for 32-bit iOS Simulator
This includes original changes authored by Orfeas in the build script to
(re)include i386 support as well as updates in the OpenSSL.xcodeproj.
However, updates in binaries and frameworks have been removed as Marcin
would probably like to perform the update himself. This commit also
includes a minor fix in the Xcode project file that puts the new
"opensslconf-i386.h" header into the 'public headers' section so that
it's actually included into the produced frameworks.

Signed-off-by: Orfeas Zafeiris <orfeaz@gmail.com>
Signed-off-by: Alexei Lozovsky <alexei@cossacklabs.com>